// apps/server/src/generated/prisma/models/CrmBlacklist.ts
export type CrmBlacklistModel = runtime.Types.Result.DefaultSelection<Prisma.$CrmBlacklistPayload>;
// ...
export type $CrmBlacklistPayload<
  ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs
> = {
  name: 'CrmBlacklist';
  objects: {
    organization: Prisma.$OrganizationPayload<ExtArgs>;
  };
  scalars: runtime.Types.Extensions.GetPayloadResult<
    {
      id: string;
      organizationId: string;
      emailHash: string;
      maskedEmail: string;
      reason: string;
      sourceAccountId: string | null;
      sourceContactId: string | null;
      sourceMessageId: string | null;
      createdById: string | null;
      createdByName: string | null;
      createdAt: Date;
      updatedAt: Date;
    },
    ExtArgs['result']['crmBlacklist']
  >;
  composites: {};
};