// apps/server/src/generated/prisma/models/CrmEmailVerificationCache.ts
export type CrmEmailVerificationCacheModel = runtime.Types.Result.DefaultSelection<Prisma.$CrmEmailVerificationCachePayload>
// ...
export type $CrmEmailVerificationCachePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "CrmEmailVerificationCache"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    emailHash: string
    maskedEmail: string
    domain: string | null
    status: string
    reason: string
    verifiedAt: Date
    expiresAt: Date
    checkedById: string | null
    checkedByName: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["crmEmailVerificationCache"]>
  composites: {}
}